refactor(app): extract settings persistence and PWA lifecycle
Move saveSettings/loadSettings into src/app/settings.ts (now taking an
AppSettings param and mutating in place so state.settings keeps its
identity for the ~31 readers) and registerServiceWorker/
setupOfflineDetection into src/app/pwa.ts (verbatim, no state coupling).
AppState.settings now uses the shared AppSettings type. main.ts:
1934 -> 1845 lines.

These paths are not e2e-covered, so verified by manual /browse smoke:
settings save/load round-trips through localStorage (toggle persists
across reload); SW correctly skips in dev; offline detection inits clean.

Verified: typecheck/lint clean, unit 223 pass, e2e 6 pass / 1 skip.
Autoreview clean.
